tmf: Bug 495219: Fix NPE in checkpoint indexer seeking on disposed trace
In BTree, if it is disposed (file is null), return -1 (insertion point
0) in the binary search. This will correspond to checkpoint 0 in the
flat array.

In FlatArray, if it is disposed (file is null) return null as checkpoint
in get(long).

In TmfCheckpointIndexer, if the returned checkpoint is null in
restoreCheckpoint(long), return the context given by the trace
implementation of seekEvent((ITmfLocation) null). It should properly
handle a disposed trace.

Change-Id: I3148d2e10118628e387d206054de55d5238a9807
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/74474
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>