lttng: Make use of "is_pic" event field in debug info analysis
A field with name "is_pic" is now present in LTTng 2.8 final in
"bin_info" events to indicate if the referred binary is PIC
(position-independent code) or not. We can make use of this
field to identify PIC binaries much more reliably than by just
looking at the file name.

Bug: 495551

Change-Id: Ib553f43e8236c399c56387ff8e20a8428c688fb5
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/74816
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI