os.linux: Add Next/Previous TID event action in CFV
Expose actions in the Control Flow View to navigate to the
previous or next trace event for a given row/thread.

Unlike the "Go to state change" action, this action will stop
on all trace events related to that thread. To do so, we need
to stop on every trace event to test if they match. Since this
operation is unbounded in terms of number of events to read,
it is deferred to a separate Job.

Change-Id: Ie0f4f7dc3f2d0c6e3f2d0111fa95f265dfcdf75f
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/75362
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>