lttng: Stabilize CriticalPathTest
To reproduce the issue:
Inside CriticalPathView.CriticalPathContentProvider.BuildThread.run()
add a Thread.sleep(5000) just before the refresh() call.

It fails because we get all the tree items of the Critical Path view
before the algorithm and the redrawing is done. Since not all of this
process is done in a job, the waitForJobs is not sufficient. Instead,
wait for the UI to update in the state we want to "assert" on.

Bug: 496588
Change-Id: I2617b017f1d94f000600cd2f25ccce98b3cfe5be
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/76153
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>