control: Moving the loglevels radio buttons over the drop down menu
Moving the loglevel and loglevel-only radio buttons over the drop down
menu instead of being underneath it. This is because most of the time
the drop down menu is going down and it hides the two radio buttons.
Those are in the enable event dialog box in the Control view under the
UST domain tab.

Change-Id: I434255ba3fd0b9e88cdd3c759c1a50cba09ea3c6
Signed-off-by: Bruno Roy <bruno.roy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/76236
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>