tmf: Fix intermittent failure in ProjectExplorerTraceActionsTest
To reproduce the issue:

In TmfTraceElement.delete, add a Thread.sleep(5000); right after the
syncExec that calls closeEditors.

Run ProjectExplorerTraceActionsTest.test4_05Delete then
test4_09BringToTop (you can create a new test method and call them in
that order).

The problem is that the delete test waits until the delete dialog and
the editor are closed but there is no guarantee that the file will
actually be deleted right away. While the file is still there, it is
possible that the next test will start and begin the process of
opening the same existing trace. But then the file gets deleted and
the trace doesn't validate properly (or other exceptions depending on
the timing of when the trace gets deleted).

We should make sure that the file is deleted before proceeding.

Bug: 497035
Change-Id: I33d0b45732bdde6061112736f2f1f7e46d55864e
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/76249
Reviewed-by: Hudson CI
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>