tmf: Bug 497038: Custom parser field names conflict with built-in tags
A new Tag enum is added as part of the custom parser definition model.

The 'tag' attribute is added to the XML definition.

For backward compatibility if the 'tag' attribute is not present when
loading a definition from XML, the Tag is determined from the 'name'
attribute.

In the internal payload data map of the custom event, the map's key is
changed from a String to an Object that can either be the Tag, a String,
or an enum constant, to ensure there is no conflict of keys.

If the tag is OTHER, the wizards should ensure that the user-entered
name does not conflict with one of the built-in tag labels.

Change-Id: If9bf65fb32d646cda4f24f28bf41941f1f5c85f6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/76257
Reviewed-by: Hudson CI