timing: show callStack segments in a density view
This creates a segment store for the callstack view and populates it.
Then the segments are used to fill a density view.

The segments are special segments with links to their children. This
effectively is a callgraph that can then be used for other analyses.

Examples of these analyses are:

* Flame graphs
* Tree maps
* Caller trees
* Call statistcs

These views will allow trace compass to work better as a profiler
on an application level.

Change-Id: I0a1e8eda192e0646dd02bc4134c672a392231208
Signed-off-by: Sonia Farrah <sonia.farrah@ericsson.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/76369
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI