lttng.control: Improve the enable logger handler
With the right-click enable logger in the Control view, it is now
possible to enable/disable the same logger with different loglevels
(having a single logger enabled multiple times with different
loglevels). Before, there was a command error because the loggers had
the same name.

Also added a work-around for an LTTng bug with the enabling and
disabling of JUL events. The bug is now fixed, and will probably be in
LTTng 2.8.1.

Change-Id: If7e5107846a2495d7835e8d1d6e0d22462a786c3
Signed-off-by: Bruno Roy <bruno.roy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/76360
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>