timing: Add Flame Graph View
The Call Graph analysis creates an aggregation tree for the call stack
view. The functions in the tree are aggregated by address and caller.

The tree nodes have links to their children and parents.

This tree is used to fill a flame graph view.

The Flame Graph is used to analyze the function call durations in a more
aggregated manner. This will show users which function call is consuming
the most time and what call chain is consuming the most time.

Change-Id: I4da99847ecef1e5246be27a2719957a568c8cae6
Signed-off-by: Sonia Farrah <sonia.farrah@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/76636
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>