tmf: Bug 497347: Custom event aspects incorrect in mixed experiments
Change the custom event aspects to use TmfContentFieldAspect with the
field name. In a mixed experiment, fields from different custom trace
types with the same name will be combined in a single aspect.

Use the base aspect for the Event type if specified as a custom parser
output.

Remove the internal column data array in custom event, as its
information is stored in the event content.

Do not store the custom event type name as a content field, as it is
already stored in the custom event type.

Change-Id: Ic8b39672f0b6f27439774262791d249e6ecf24b1
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/76640
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>