tmf.xml: Add mapping group to XML description
The mapping groups act as maps. They are used to avoid large if-else
condition in the XML. For example, if the relevant event for an FSM have
an integer field 'mode' and this value could be map to string. We could
save the 'mode' value in the state system and add a mapping group in the
XML to be able to directly use the string instead of the 'mode' value
when creating segments.

example of mode could be ;
* 1 - read only
* 2 - read - write
* 3 - write - only
* ....

Change-Id: I0237e62b304f984afcf78b09fd7b45c5687a80d5
Signed-off-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/76780
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>