lttng.ust: Do not block in UstDebugInfoAnalysisModule
Query methods in the module should not call waitForCompletion()
on the analysis, but should return whatever data they have
(or null).

This is because this is called by the aspects, which are called
by the event table, which are called by the UI thread! The UI
should not block while the analysis is waiting construction.

If a specific caller wants to wait for the full data, they can
call waitForCompletion() manually, like the tests do.

Change-Id: I2cf4e3544a8eea71a65f1e16fe7d4b5ba881d5ca
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
(cherry-pick of 599e5482448fede7d30a73bd9a3471b97be7ba4f)
Reviewed-on: https://git.eclipse.org/r/80340
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>