tmf: Support default timestamp output format in custom parsers
The timestamp output format can now be left blank to use the default
Time Format preference. In that case the Timestamp base aspect will be
used.

When the default format is used, the Timestamp is no longer stored as a
content field, as it is already stored in the event's timestamp. 

Change-Id: I1e6f83ffb35f44be7828e3d25ee43f87fa8a97e1
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/77497
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>