tmf: Remove function name from ITmfCallsite
It just so happens today that the UST Debug-info analysis makes
use of addr2line to retrieve the source code location AND the
function names, but those two are completely separate concepts.

Since "function name" is not relevant for source lookup, it
should be removed from the concept of what is a call site.
ISymbolProvider should be the sole provider of function names,
especially now that it has been decoupled from ITmfCallsite.

While at it, add a new version of #getLineNumber() which returns
a "@Nullable Long", since it's very possible for callsites to
have a file name but no line number available.

Change-Id: I9b7407876166d9cc1e174804145937388306137d
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/77554
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>