linux/lttng: Add a swtbot test for UI responsiveness of kernel views
This adds abstract swtbot classes to open a trace, then navigate it
(full range, zooms, scroll, zoom back out) while some views are opened. This
test is not meant to be run in the performance test suite, but enabling JUL
logging with it allows to have a reproducible series of UI operations and
analyze the so-obtained traces.

Change-Id: Ia43c6b1244b989a6888547b43247bc9b30435dee
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/78271
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI