linux.core: Support IPIs in kernel analysis (Bug 498215)
Interprocessor interrupts (IPIs) are special types of interrupts by
which one processor will interrupt another in a multi-core system. They
are typically used for

* cache flushes
* shutdowns

They are not logged with standard events, but rather events looking like
"x86_irq_vectors_thermal_apic_exit".

This patch makes them register as IRQs (which they are) in the control
flow view and resources view.

Change-Id: I5380b6bc2d18f2ee98f7675eb1918794a78a71b3
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/77731
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-by: Francis Giraldeau <francis.giraldeau@gmail.com>
Tested-by: Francis Giraldeau <francis.giraldeau@gmail.com>