tmf/ust: deprecate AbstractCallStackAnalysis and move the UST module to core
The only reason the CallStackAnalysisModule were in the UI was to register the
view output to it. A listener now does that part so the class in the tmf.ui
plugin can be deprecated and LttngUstCallStackAnalysis can also be moved to
the .core plugin and be made internal while at it.

Change-Id: I80b4c34ff35855b5261a1270e67d10df865d6e0d
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/77781
Reviewed-by: Hudson CI