linux.ui: extract control flow view optimizer
The control flow view optimizer algorithm is now extracted into
a Function<>, this allows for better testing of the code and
improves modularity.

To change the optimizer behaviour, one needs to override the
getUpdatedSchedulingColumn() function to return another algorithm.
Consider ILinkEvents to be edges in a graph where the vertices are
the entries.

Change-Id: I2a8cb686e2c31589873006d1fca2f489ad724b33
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/77871
Reviewed-by: Hudson CI
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>