custom charts: Add interfaces for computing data
This patch adds three interfaces that are used for
consuming objects in the source of the data provider.
An object is first consumed by a chart consumer. This
first consumer sends the object to all of its series
consumers. Then, each series consumer passes the
object to individual data consumer.

While this may seems a lot, it is rather simple and
it allows going to the stream of data only once. It
also provides flexible data computing as well as
reusing code for different charts.

Change-Id: Ife9f321fa840cde6f499a09094ec996d6bf669f2
Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabrielpolloguilbert@gmail.com>
Reviewed-on: https://git.eclipse.org/r/78011
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI