lttng.control: Fix enabling logger by specifying the name
If a user enters a specific logger name to be enabled, we should not
enable all loggers. With the original code, if there was not a logging
application running the fIsAllLoggers flag would evaluate true, even if
the user entered a specific name in the text field (which I don't think
is the desired behavior).

Change-Id: Iab37f45dedba82b2c0d420432488e7fd7c770511
Signed-off-by: Bruno Roy <bruno.roy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/78641
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>