tmf.ui: fix potential priority inversion in symbol providers.
It is a common practice to set a priorty of -Long.MIN_VALUE to an item to
make sure it is last. With the current implementation, it will always be
first, appearing before even Long.MAX_VALUE.

This patch fixes that issue.

Also, the constructor could fail out with an exception in several paths.
Now this is done by the factory method. This avoids having partial
objects in the heap which can be a security issue.

Change-Id: I392205a0b1a0557c168a2ae547b57915ee57f6f3
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/78758
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Tested-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-by: Hudson CI