tmf: Fix lingering zoom range markers on mouse release
When zooming with the right mouse button, when the button is released
the drag state should be reset before the change of window range redraws
so that the zoom drag selection lines and background are not drawn one
last time.

The redraw used to trigger the paint event after a short delay (in the
platform) at which time the drag state had been reset, but a recent
change in the time graph viewer has forced the paint event to happen
immediately by a call to Control.update().

Change-Id: Ifd60e2053bd7736ca37749cb1b299ea0d4852d79
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/79552
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>