flamegraph: fix invalid window range
When loading a flamegraph we need to reset the start
and finish time so that the time range is fixed and the
whole flame graph range is displayed.

Otherwise, we end up showing a range of 1 ns after
zooming once and switching the trace.

Change-Id: I85e950754e1361c7fe31739a280a3b14152751aa
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/79633
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
(cherry picked from commit c418199f8941c3c3ba36b2ad995b6244146d1f1d)
Reviewed-on: https://git.eclipse.org/r/79662