analysis: make call graph analysis non automatic
Having automatic it will be executed for every trace since it is
available for each trace type at the moment. Moreover, since it
is dependent on the call stack analysis the call call stack
analysis is also triggered. 

With this patch it is only done when one of the applicable call
graph views, e.g. Flame Graph view, is opened.

Change-Id: Id80782dc2050f8816b1440c7a9e7ff19503e0373
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/79760
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>