Add an integration tests plugin
This plugin will contain tests that possibly require many "leaf" dependencies
and that are not unit tests but instead fuller scenarios. For example,
it might test importing UST and Kernel traces together in one operation
and test for name clashes. Therefore it is not a unit test for UST or
Kernel.

The plugin is not added in the parent pom.xml for now as it would fail
because it does not have a test yet.

Change-Id: I6f3af925700b78adbca04dbdf4017f59581ae9e4
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/79789
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>