tmf: Bug 500542: Missing zoomed events in time graph entry
When adding an event that starts before the current zoomed event list's
first event, the list is deemed to be incomplete and missing events at
the beginning of the zoom range. The list is cleared and the new event
is added. The remaining events for the full zoom range should follow.

Change-Id: I7057fbaac6ca3a57b2923c786c08142b2ebf6fc3
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/80076
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>