timing: Bug 500592: fix symbols for with debug info of Lttng 2.8
The process ID of the trace application is propagated so that it can
be used for the Function Duration Density view and Flame Graph view.

For the Flame Graph it uses the the timestamp of the maximum segment to
get the symbol name from the LTTng UST symbol provider.

For the Function density view it takes the start time of the segment.

Change-Id: Id3eb817f156e30b9be4996b732de05bc335418b9
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/80298
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit c2845a63b7bf74179d9a7b8207dbb8a68aa716d8)
Reviewed-on: https://git.eclipse.org/r/80486