os.linux: Create PPID and SYSTEM_CALL attributes on demand
This patch ensures that threads' PPID and SYSTEM_CALL attributes
are only created when the state is changed instead of on a
sched_switch. The methods which require said attributes are adapted
to handle correctly the absence of said attributes.
This results in a state system with 0 - 40% less attributes, thus
reducing the legacy SHT's space on disk by the same size.

Change-Id: I1461824623912f8d2ea32bb2c97f0398b08f9aea
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-on: https://git.eclipse.org/r/80849
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>