tmf: Fix hang in TarInputStream.skipToEntry when importing large tar.gz
When importing a large tar.gz archive, it can hang in
TarInputStream.skipToEntry. Depending on the size of the entries in the
Tar and the order in which they are read, it is possible that bytesread
will overflow and become negative. Because of that, the code that skips
bytes fails to skip properly and enters in an infinite loop.

This patch fixes the issue by making the bytesread a 'long'. Because
this class is in the platform, several files had to be copied over in
order to get the fix right away. Those files will be removed once the
fix is in the platform and that the fix is released.

Bug: 501379
Change-Id: Ia1613da08715968f23a803519a2e78aa7c51482f
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/81053
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>