linux.core: add Tux Icon to linux core analysis
According to the site : http://isc.tamu.edu/~lewing/linux/

Here are the results of an idea acquired from discussions
on the linux-kernel mailing list, and an initial suggestion
by Alan Cox.

Feel free to do whatever you see fit with the images, you
are encouraged to integrate them into other designs that
fit your need. Comments suggestions are also welcome, so
please tell me what you think of these. I suggest that you
look at some of the other images available with integrated
text.

The backgrounds of these images are random colors (if your
viewer doesn't support transparent gifs). This is because
I want to be able to keep the outline clean (except when
blending into a scene or title bar). Each in-line image is
now also a link to the corresponding gif so that they are
more easily retrieved. The images I actually work from are
tifs which I'll make available if there is interest.

Neal Tucker was kind enough to provide a scalable vector
based postscript version of the black and white penguin.

Permission to use and/or modify this image is granted provided
you acknowledge me lewing@isc.tamu.edu and The GIMP if someone
asks.

This patch will still require a CQ, but it looks safe to push
to gerrit.

CQ: https://dev.eclipse.org/ipzilla/show_bug.cgi?id=12011

Change-Id: Ia31cecaf6d2e870e37e8425dca475641c23d90cf
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/81654
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>