tmf.ui: bug 503678 The ss explorer should not execute analyses
And because the state system does not execute the analyses, the state
systems may not exist when the tree is built. The hierarchy is thus
changed so that the analyses themselves appear as top-level items and
their children are the state system. This way, if the state system did
not exist at first and appears later, it will show when the view
refreshes (without having to refresh the whole view).

Change-Id: Icc750abcdfb27d74d486b9fdcf8b3e910310684f
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/83811
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>