lttng: Fix ControlViewTest
The regex replacement of the user home in command output are quoted to
avoid stripping backslashes in a Windows path.

The list provider regex patterns are updated to include the new command
options '-j' '-l' and '-p'.

The target node state is only set to CONNECTED after the node
configuration jobs are completed.

The SWTBot test waits for the node configuration jobs to be completed
before continuing with the test scenario.

The UI test is changed to use a wait condition instead of a loop with
delay when waiting for the target node state to be connected.

Change-Id: Ie05f58181119cf46d8cdacb8313856f6b3c4da30
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/82972
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>