lttng: Fix test when in the presence of long workspace path
LTTngServiceMiTest validated against the old MI XSD which doesn't
allow sessions paths longer than 255 characters. When the workspace
of the test is too long, this fails. The test doesn't really need
a variable and can be hard-coded to any path. So this patch removes
the use of the workspace variable for this case.

See also bug 490557.

Change-Id: I382da2449c485e828c62d4ef73e3baf75e438374
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/83737
Reviewed-by: Hudson CI
Reviewed-by: Jonathan Rajotte Julien <jonathan.rajotte-julien@efficios.com>
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>