tmf: Deprecate TimeGraphCombo and support columns in TimeGraphViewer
The class TimeGraphCombo is deprecated and AbstractTimeGraphView method
getTimeGraphCombo() is deprecated and returns null.

The TimeGraphViewer supports columns and label provider with new methods
setTimeGraphLabelProvider(), setColumns(String[]), setWeights(int[]) and
getTree(). The tree contains no tree items and is only used to show
column headers and allow for column handling.

If no column names are specified, a single nameless column is used.

If no label provider is specified, the time graph entry's name and the
time graph presentation provider's item image is shown in the first
column.

The TimeGraphViewer is made to extend the Viewer class.

The TimeGraphControl supports auto resize of columns to grow columns at
the item text's preferred width. Auto resize turns off automatically the
first time that a user resizes a column to be narrower than its
preferred width.

The tree expand and collapse controls are changed to be triangle widgets
that are bound in size by the item's height.

The name space split line is made easier to hover with the mouse.
Double-clicking it snaps the name space to the right of the last visible
column text's preferred width. The split line can now be dragged even if
the viewer has no input model.

The TimeGraphControl methods getTree(), getLabelProvider() and
getItemBounds(ITimeGraphEntry) are added to help with UI testing.

Change-Id: I1061371b083499298da4f7c5640929e63ed651db
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/84135
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>