tmf: Bug 496504: Fix duplicate child entries in Control Flow view
This could happen while building the state system if a child process is
first discovered in an early iteration and its parent process is found
in a second iteration. Removing the child from its old parent trace
entry would break the link to the new parent that was just set. On a
third iteration it would then add the child to the same parent again.

The TimeGraphEntry is changed to be more robust in its handling of the
parent-child relations.

Change-Id: I99743d80484d335180459df144067fc4494b8052
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/83961
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>