tmf: Introduce "New view" action for views based on TmfView
All views based on TmfView can now optionally have a new action button
in their view menu providing an easy way to spawn a view of the same
type. The interface ITmfAllowMultiple is introduced for this purpose.

Introduce a TmfViewFactory Singleton.

This is an initial work toward providing a cloning ability to
TmfView-based views.

Change-Id: Ie8a0c4af305edb6f14968188cfd8feda8983a8cd
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/84903
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>