tmf: Make AbstractTimeGraphView implement ITmfPinnable
When pinned the time graph view ignores trace change signals. The view
tab label is changed to indicate the pinned trace. When receiving trace
closed signal for the trace to which the view is currently pinned, the
view is unpinned.

When receiving window range and selection range update signals, the time
graph view ignores the range received in the signal and instead
synchronizes itself with the range stored in the trace manager.

Change-Id: I6bf2329f0f932feba20f06c971a1adb4de940b6c
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/84900
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI