tmf: Fix MenuDetect handling in TimeGraphControl
- If the menu is invoked with the mouse over the name space, then there
is no need to delay the menu detect handling until mouseUp, since the
right mouse button isn't being used to zoom

- If it is a pending menu detect callback on mouseUp, and another drag
operation is in progress or it is a drag zoom outside of the margin,
then make sure the menu is not made visible by setting e.doit to false.

- If the menu is invoked with the mouse over the name space, or with the
keyboard, if there is any drag operation in progress make sure the menu
is not made visible by setting e.doit to false.

- If the menu detect is handled immediately (not after a pending menu
detect on mouseUp) then there is no need to set the menu visible
programmatically since the platform will do it after the MenuDetect
event handling when e.doit is true.

Change-Id: I0d76b9908782fc853a065567a6e907da5e2d9c86
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/84434
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>