tmf.ui: make TimeEvents in ControlFlowView have thickness
This makes active threads much easier to spot.
Adds notion of thickness to the "StateItem". In order to
use it, override StateItem#getHeightFactor().

The patch also introduces the notion of a style map.
This is loosely based on CSS fill styles. The styles
in the map are suggestions and do not need to be implemented
by the UI, but could at a later date.

The styleMap is passes as follows where each step can override the
base map:

* LinuxStyle (Style on an analysis level)
* StateItem (Style on a per-view level)
* PresentationProvider (Style on a per-event level)

Change-Id: I0a5f33d958a1ec1746ace9a66fecd728fb5a68ed
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/85092
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>