os.linux: Model the CPU run queue of each thread
Add a new attribute to the kernel state system:

THREADS
 +--1000
     +--Current_cpu_rq

Which will track on which CPU's run queue this thread
is currently located. Unlike the cpu -> Current_thread
attribute, several threads can be present on the same
CPU's run queue at the same time.

This will allow for example tracking on which CPU a
thread is expecting to go when it is in the WAIT_FOR_CPU
state.

This requires tweaks to the sched_switch, sched_wakeup
and process_fork handlers, as well as adding a new handler
for the sched_migrate_task event (which simply moves a
non-running thread from one run queue to another).

Change-Id: I6e0386214df195eb6389cec87f2e6a42fdc0691d
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/85805
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>