os.linux: use Current_cpu_rq for control flow view tool tip
Enable the displaying on hovering of the CPU run queue information for
thread in the WAIT_FOR_CPU state.

Using Current_cpu_rq is also valid for currently running thread.

Move the addition of the CPU tool tip attribute to
getEventHoverToolTipInfo so the presented information reflect the
reality at the hovering time.

Change-Id: Ib01cfd883e5ebc4b3f24143bebaf0ad5da74ff69
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Reviewed-on: https://git.eclipse.org/r/85809
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>