analysis.lami: Fix internal signaling with several views on the same report
Use the parent LAMI tab page instance as the key for internal signaling.

The signal key is designed to restrict the signal to interested objects.
It was based on the hash of the immutable fResultTable field. Since for multiple
view of the same report the fResultTable is the same object, signals were
processed by objects that were not concerned. Using the result table
instance/hash as a key is not sufficient.

Bug: 495322

Change-Id: Ifba2056b10cd254d1a8c6c030c4b7263463ac834
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
(cherry picked from commit 3f506e2542817000a0851dbd0bfaf5683ed04b5b)
Reviewed-on: https://git.eclipse.org/r/86171
Reviewed-by: Hudson CI