linux.core: fix memory allocation issue.
The order field in page alloc/free say how many pages are allocated.
It was ignored, which gave false totals.

It is artificially limited to 2^62 page size, as larger would garanty
an overflow.

With this patch, memory overflows can occur much faster, but only on
systems which trace or report over Long.MAX_VALUE memory.

Change-Id: I5c805115cb3aaa5c6bc6146a6dc0db20f11854ef
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/87181
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>