swtbot: Fix time graph item bounds and menu detect event coordinates
The method TimeGraphControl.getItemBounds(ITimeGraphEntry) should return
bounds relative to the item's parent, e.g. the time graph control, not
its parent.

The method SWTBotTimeGraphEntry.contextMenu(Control) should fire a
SWT.MenuDetect event where the coordinates are relative to the display.
The item should also be selected after firing the SWT.MenuDetect event
but before firing the SWT.Show menu event.

The method SWTBotTimeGraphEntry.select() should replace the selection
with the selected item, not add it to the selection. It should also fire
a SelectionEvent and a SelectionChangedEvent.

Change-Id: I5743af433c51b97255aafb0ce433d69069858e50
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/87498
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>