common.core: clean up ProcessUtils a bit
While the code does work, it has some issues that could potentially be
problems with the code if not used in the way it should.

* Extract magic number to constants

This is simple, it makes modifying the behaviour easier.

* Explicitly fill code blocks for catch blocks

Improve code readability. It makes it clear that it is intentional.

* Remove redundant modifier on interfaces

An interface is always static, there is no need to add the keyword
"static" to it.

* Add charsets to inputstream

This explicitly expects the default encoding (UTF8). However, this is
not the ideal solution, it would be much better, later to have an
interface with the charset passed in it. It can be used if we want to
launch a process in another machine.

Change-Id: I5ef919e724dc28d9e73c5276811af88b4de12f67
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/88407
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Tested-by: Alexandre Montplaisir <alexmonthy@efficios.com>