timing: ss statistics use the common statistics
This patch has the SegmentStoreStatistics class use a Statistics object
instead of calculating the statistics themselves (it will later be
deprecated entirely, but there are too many things to fix for one patch)

It deprecates the methods of the AbstractSegmentStoreAnalysis that used
the SegmentStoreStatistics class.

It also deprecates the current abstract statistics view and viewer in
the UI and replaces them with one using the new statistics class.

Change-Id: Idfdeed3198d70d0eae7b75b90453f67ddb138809
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/88901
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>